book_reader/web_browser/requirements.txt
msedge-selenium-tools
